🔧 use ENV variables to define lsf config params
