+++ toil_container/lsf.py
@@ -3,51 +3,19 @@
 
 from collections import defaultdict
 from datetime import datetime
-import base64
-import json
 import os
-import random
 import subprocess
-import time
 
 from toil.batchSystems.lsf import LSFBatchSystem, logger
 from toil.batchSystems.abstractBatchSystem import UpdatedBatchJobInfo
 
-
-from toil_container.lsf_helper import parse_memory
-from toil_container.lsf_helper import PER_CORE
-
-
-_RESOURCES_START_TAG = "__rsrc"
-_RESOURCES_CLOSE_TAG = "rsrc__"
-
-try:
-    _MAX_MEMORY = int(os.getenv("TOIL_CONTAINER_RETRY_MEM", "60")) * 1e9
-    _MAX_RUNTIME = int(os.getenv("TOIL_CONTAINER_RETRY_RUNTIME", "40000"))
-except ValueError:  # pragma: no cover
-    _MAX_MEMORY = 60 * 1e9
-    _MAX_RUNTIME = 40000
-    logger.error("Failed to parse default values for resource retry.")
-
-
-def with_retries(operation, *args, **kwargs):
-    """Add a random sleep after each retry."""
-    latest_err = Exception
-
-    for i in [2, 3, 5, 10]:
-        try:
-            return operation(*args, **kwargs)
-        except subprocess.CalledProcessError as err:
-            time.sleep(i + random.uniform(-i, i))
-            latest_err = err
-            logger.error(
-                "Operation %s failed with code %d: %s",
-                operation,
-                err.returncode,
-                err.output,
-            )
-
-    raise latest_err  # pragma: no cover
+from toil_container.lsf_helper import (
+    MAX_MEMORY,
+    MAX_RUNTIME,
+    build_bsub_line,
+    decode_dict,
+    with_retries,
+)
 
 
 class CustomLSFBatchSystem(LSFBatchSystem):
@@ -100,7 +68,7 @@ class CustomLSFBatchSystem(LSFBatchSystem):
 
             try:  # try to update runtime if not provided
                 jobNode = self.boss.Id2Node[jobID]
-                runtime = runtime or _decode_dict(jobNode.unitName).get("runtime", None)
+                runtime = runtime or decode_dict(jobNode.unitName).get("runtime", None)
                 jobname = f"{env_jobname} {jobNode.jobName} {jobID}"
             except KeyError:
                 jobname = f"{env_jobname} {jobID}"
@@ -235,8 +203,8 @@ class CustomLSFBatchSystem(LSFBatchSystem):
 
             retry_type = "memlimit" if term_memlimit else "runlimit"
             jobNode.jobName = (jobNode.jobName or "") + " resource retry " + retry_type
-            memory = _MAX_MEMORY if term_memlimit else jobNode.memory
-            runtime = _MAX_RUNTIME if term_runlimit else None
+            memory = MAX_MEMORY if term_memlimit else jobNode.memory
+            runtime = MAX_RUNTIME if term_runlimit else None
             bsub_line = self.prepareBsub(jobNode.cores, memory, jobID, runtime)
 
             if retry_type not in self.boss.resourceRetryCount[jobID]:
@@ -259,80 +227,3 @@ class CustomLSFBatchSystem(LSFBatchSystem):
                 .strip()
                 .split("\n")[1:]
             }
-
-
-def build_bsub_line(cpu, mem, runtime, jobname, stdoutfile=None, stderrfile=None):
-    """
-    Build an args list for a bsub submission.
-
-    Arguments:
-        cpu (int): number of cores needed.
-        mem (float): number of bytes of memory needed.
-        runtime (int): estimated run time for the job in minutes.
-        jobname (str): the job name.
-        stdoutfile (str): filename to direct job stdout
-        stderrfile (str): filename to direct job stderr
-
-    Returns:
-        list: bsub command.
-    """
-    bsubline = [
-        "bsub",
-        "-cwd",
-        ".",
-        "-o",
-        stdoutfile or "/dev/null",
-        "-e",
-        stderrfile or "/dev/null",
-        "-J",
-        f"'{jobname}'",
-    ]
-    cpu = int(cpu) or 1
-    if mem:
-        mem = float(mem) / 1024 ** 3
-        logger.warning(f"PREPARE BSUB: GET PER CORE RESERVATION: {PER_CORE}")
-        if PER_CORE:
-            mem = mem / cpu
-
-        mem = mem if mem >= 1 else 1.0
-
-        mem_resource = parse_memory(mem)
-        mem_limit = parse_memory(mem)
-
-        bsubline += ["-R", f"select[mem>{mem_resource}]"]
-        bsubline += ["-R", f"rusage[mem={mem_resource}]"]
-        bsubline += ["-M", str(mem_limit)]
-
-    if cpu:
-        bsubline += ["-n", str(cpu)]
-
-    if runtime:
-        bsubline += [os.getenv("TOIL_CONTAINER_RUNTIME_FLAG", "-W"), str(int(runtime))]
-
-    if os.getenv("TOIL_LSF_ARGS"):
-        bsubline.extend(os.getenv("TOIL_LSF_ARGS").split())
-
-    # log to lsf
-    logger.info("Submitting to LSF with: %s", " ".join(bsubline))
-    return bsubline
-
-
-def _encode_dict(dictionary):
-    """Encode `dictionary` in string."""
-    if dictionary:
-        return (
-            f"{_RESOURCES_START_TAG}"
-            f"{base64.b64encode(json.dumps(dictionary).encode()).decode()}"
-            f"{_RESOURCES_CLOSE_TAG}"
-        )
-    return ""
-
-
-def _decode_dict(string):
-    """Get dictionary encoded in `string` by `_encode_dict`."""
-    if isinstance(string, str):
-        split = string.split(_RESOURCES_START_TAG, 1)[-1]
-        split = split.split(_RESOURCES_CLOSE_TAG, 1)
-        if len(split) == 2:
-            return json.loads(base64.b64decode(split[0]))
-    return {}
